// src/app/shared/models/group-metric.model.ts
export class GroupMetricModel {
	id: string;
	createdAtUTC: string;
	updatedAtUTC: string;
	groupId: string;
	metricForHourUTC: string;
	metricForHourUTCStartTick: number;
	metricForHourUTCHour: number;
	metricForHourUTCDay: number;
	metricForHourUTCWeek: number;
	metricForHourUTCMonth: number;
	metricForHourUTCYear: number;
	numComments: number;
	numPosts: number;
	numCommentsPerPost: number;
	numReactions: number;
	numShares: number;
	numTextPosts: number;
	numVideoPosts: number;
	numPhotoPosts: number;
	numLinkPosts: number;
	numEventPosts: number;
	memberCount: number;
	memberRequestCount: number;
	numCommentsOnVideoPosts: number;
	numCommentsOnPhotoPosts: number;
	numCommentsOnLinkPosts: number;
	numCommentsOnEventPosts: number;
	numCommentsOnTextPosts: number;
	numReactionsOnTextPosts: number;
	numReactionsOnVideoPosts: number;
	numReactionsOnPhotoPosts: number;
	numReactionsOnLinkPosts: number;
	numReactionsOnEventPosts: number;
	numSharesOnTextPosts: number;
	numSharesOnVideoPosts: number;
	numSharesOnPhotoPosts: number;
	numSharesOnLinkPosts: number;
	numSharesOnEventPosts: number;
	numWeekOfMonth: number;
}